+++ game_defence/src/modules/camera/helpers.hpp
@@ -9,7 +9,7 @@ namespace game2d {
 [[nodiscard]] glm::vec2
 mouse_position_in_worldspace(entt::registry& r);
 
-[[nodiscard]] glm::ivec2
+[[nodiscard]] glm::vec2
 position_in_worldspace(entt::registry& r, const glm::ivec2& pos);
 
 } // namespace game2d
